pyproject.toml
[project]
name = "ecommerce-negotiator"
version = "0.1.0"
description = "AI-powered ecommerce negotiation agent"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.3",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "httpx>=0.26.0",
    "aiosqlite>=0.19.0",
    "playwright>=1.41.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
